- Add wallet/shared package - bring account response type into shared package
