+++ package.json
@@ -47,11 +47,11 @@
     "typescript": "^5.4.5"
   },
   "engines": {
-    "pnpm": "^9.1.1",
+    "pnpm": "^9.1.3",
     "npm": "pnpm",
     "yarn": "pnpm",
     "node": "^20.12.1"
   },
   "private": true,
-  "packageManager": "pnpm@9.1.1"
+  "packageManager": "pnpm@9.1.3"
 }
