+++ packages/wallet/frontend/src/components/dialogs/FundAccountDialog.tsx
@@ -3,7 +3,7 @@ import { Dialog, Transition } from '@headlessui/react'
 import { Fragment } from 'react'
 import { Input } from '@/ui/forms/Input'
 import { Button } from '@/ui/Button'
-import { Account, accountService, fundAccountSchema } from '@/lib/api/account'
+import { Account, accountService } from '@/lib/api/account'
 import { useDialog } from '@/lib/hooks/useDialog'
 import { ErrorDialog } from './ErrorDialog'
 import { getCurrencySymbol, getObjectKeys } from '@/utils/helpers'
@@ -11,6 +11,7 @@ import { useZodForm } from '@/lib/hooks/useZodForm'
 import { Form } from '@/ui/forms/Form'
 import { useRouter } from 'next/router'
 import { useOnboardingContext } from '@/lib/context/onboarding'
+import { fundAccountSchema } from '@wallet/shared/src/responses'
 
 type FundAccountDialogProps = Pick<DialogProps, 'onClose'> & {
   account: Account
